リポジトリに変換処理移行
